Move mailer to server-side; sample erb template
