+++ www/roster/templates/ldap_notification.erb
@@ -1,0 +1,9 @@
+The current roster for this group can be found at
+
+  https://whimsy.apache.org/roster/<%= type %>/<%= name %>
+  
+LDAP update details:
+
+  <%= details %>
+
+This message was generated from Whimsy.
